Refactor MessagesAdapter
